将用例进行拆分，从而与c文件一一对应
